real_empire_studio/requirements/dev.txt
-r base.txt

django-extensions==1.5.7
django_debug_toolbar==1.4
Werkzeug==0.10.4
